refactor: 시도 횟수 관련 기능 클래스 분리
