+++ src/test/java/racingcar/controller/RaceControllerTest.java
@@ -39,49 +39,6 @@ class RaceControllerTest extends NsTest {
         assertThat( raceController.selectWinners(cars)).isEqualTo(List.of("pobi", "jun"));
     }
 
-    @Test
-    @DisplayName("반복 횟수에 숫자가 아닌 것을 입력하면 예외가 발생한다")
-    void playCountFormatError(){
-        assertSimpleTest(()-> {
-            run("삼");
-                    assertThatThrownBy(() -> raceController.validatePlayCount())
-                            .isInstanceOf(IllegalArgumentException.class);
-                }
-        );
-    }
-
-    @Test
-    @DisplayName("반복 횟수에 숫자를 입력하면 예외가 발생하지 않는다")
-    void playCountValidFormat(){
-        assertSimpleTest(()-> {
-            run("5");
-                    assertThatCode(() -> raceController.validatePlayCount())
-                            .doesNotThrowAnyException();
-                }
-        );
-    }
-    @Test
-    @DisplayName("반복 횟수에 0 이하이거나 10 초과 값을 입력하면 예외가 발생한다")
-    void playCountLimitError(){
-        assertSimpleTest(()-> {
-            run("0");
-                    assertThatThrownBy(() -> raceController.validatePlayCount())
-                            .isInstanceOf(IllegalArgumentException.class);
-                }
-        );
-    }
-
-    @Test
-    @DisplayName("반복 횟수에 1에서 9사이의 숫자를 입력하면 예외가 발생하지 않는다")
-    void playCountValidLimit(){
-        assertSimpleTest(()-> {
-            run("5");
-                    assertThatCode(() -> raceController.validatePlayCount())
-                            .doesNotThrowAnyException();
-                }
-        );
-    }
-
     @Override
     protected void runMain() {
